requirements.txt
streamlit>=1.31.0
openai>=1.12.0
sentence-transformers>=2.2.2
pinecone-client>=3.0.0
python-dotenv>=1.0.0
pypdf>=4.0.0
